Springboot upgrade test fixes
