+++ build.gradle
@@ -231,6 +231,16 @@ dependencyCheck {
   }
 }
 
+configurations.all {
+  resolutionStrategy {
+    eachDependency { DependencyResolveDetails details ->
+      if (details.requested.group == 'org.hibernate.orm' && details.requested.name == 'hibernate-core') {
+        details.useVersion '6.6.1.Final'
+      }
+    }
+  }
+}
+
 dependencyManagement {
   dependencies {
     dependency group: 'commons-io', name: 'commons-io', version: '2.17.0'
@@ -286,7 +296,7 @@ def versions = [
   junit         : '5.11.3',
   junitPlatform : '1.11.3',
   reformLogging : '6.1.7',
-  springDoc     : '1.8.0',
+  springDoc     : '2.2.0',
   serenity      : '3.4.2',
   gradlePitest  : '1.15.0',
   pitest        : '1.16.1',
@@ -295,8 +305,7 @@ def versions = [
   logbook       : '3.9.0',
   tomcat        : '10.1.28',
   testcontainers: '1.20.3',
-  mapStruct     : '1.3.0.Final',
-  logbook      : '3.9.0'
+  mapStruct     : '1.3.0.Final'
 ]
 
 ext.libraries = [
@@ -314,6 +323,7 @@ ext['snakeyaml.version'] = '2.0'
 dependencies {
   implementation group: 'org.mapstruct', name: 'mapstruct-jdk8', version: versions.mapStruct
   implementation group: 'org.mapstruct', name: 'mapstruct-processor', version: versions.mapStruct
+  implementation 'org.junit.jupiter:junit-jupiter:5.8.1'
   annotationProcessor group: 'org.mapstruct', name: 'mapstruct-processor', version: versions.mapStruct
 
   testImplementation group: 'org.mapstruct', name: 'mapstruct-jdk8', version: versions.mapStruct
@@ -327,13 +337,16 @@ dependencies {
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa'
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter-cache'
 
+  implementation group: 'org.springframework.boot', name: 'spring-boot-starter-validation', version: '3.3.4'
+
+
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter-oauth2-client', version: '2.7.15'
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter-oauth2-resource-server'
   implementation group: 'org.springframework.cloud', name: 'spring-cloud-context', version: '4.1.4'
 
   implementation group: 'org.springframework.cloud', name: 'spring-cloud-starter-openfeign', version: '4.1.3'
 
-  implementation group: 'org.springdoc', name: 'springdoc-openapi-starter-webmvc-ui', version: '2.2.0'
+  implementation group: 'org.springdoc', name: 'springdoc-openapi-starter-webmvc-ui', version: '2.6.0'
 
   implementation group: 'org.apache.httpcomponents.client5', name: 'httpclient5', version: '5.3.1'
 
@@ -341,12 +354,10 @@ dependencies {
 
   implementation('org.springframework.retry:spring-retry')
 
-  implementation group: 'org.springdoc', name: 'springdoc-openapi-ui', version: versions.springDoc
-
   implementation group: 'org.zalando', name: 'logbook-core', version: versions.logbook
   implementation group: 'org.zalando', name: 'logbook-spring-boot-starter', version: versions.logbook
   implementation group: 'org.zalando', name: 'logbook-httpclient5', version: versions.logbook
-  implementation group: 'org.zalando', name: 'problem-spring-web-starter', version: '0.26.2'
+  implementation group: 'org.zalando', name: 'problem-spring-web-starter', version: '0.29.1'
 
   implementation group: 'net.minidev', name: 'json-smart', version: '2.4.10'
 
@@ -356,7 +367,7 @@ dependencies {
 
   implementation group: 'com.github.hmcts.java-logging', name: 'logging', version: versions.reformLogging
 
-  implementation group: 'org.hibernate', name: 'hibernate-core', version: '6.6.0.Final'
+  implementation group: 'org.hibernate', name: 'hibernate-core', version: '6.6.1.Final'
   implementation group: 'org.hibernate', name: 'hibernate-validator', version: '8.0.1.Final'
 
   implementation group: 'org.apache.logging.log4j', name: 'log4j-api', version: '2.24.1'
@@ -399,9 +410,20 @@ dependencies {
     exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
   }
 
-  testImplementation(group: 'com.github.tomakehurst', name: 'wiremock-jre8', version: '2.35.2') {
-    exclude group: 'com.github.jknack.handlebars.java'
+  testImplementation(group: 'com.github.tomakehurst', name: 'wiremock-jre8', version: '3.0.1') {
+    exclude group: 'org.eclipse.jetty', module: 'jetty-server'
+    exclude group: 'org.eclipse.jetty', module: 'jetty-http'
+    exclude group: 'org.eclipse.jetty', module: 'jetty-util'
+    exclude group: 'org.eclipse.jetty', module: 'jetty-io'
+    exclude group: 'org.eclipse.jetty', module: 'jetty-security'
+    exclude group: 'org.eclipse.jetty.http2', module: 'http2-server'
   }
+  testImplementation 'org.eclipse.jetty:jetty-server:11.0.15'
+  testImplementation 'org.eclipse.jetty:jetty-http:11.0.15'
+  testImplementation 'org.eclipse.jetty:jetty-util:11.0.15'
+  testImplementation 'org.eclipse.jetty:jetty-io:11.0.15'
+  testImplementation 'org.eclipse.jetty:jetty-security:11.0.15'
+  testImplementation 'org.eclipse.jetty.http2:http2-server:11.0.15'
 
   testImplementation group: 'com.github.hmcts', name: 'document-management-client', version: '7.0.1'
