+++ src/functionalTest/java/uk/gov/hmcts/reform/wataskmanagementapi/watasks/controllers/GetWorkTypesControllerTest.java
@@ -49,7 +49,7 @@ public class GetWorkTypesControllerTest extends SpringBootFunctionalBaseTest {
             "tribunal-caseworker");
 
         Response result = restApiActions.get(
-            ENDPOINT_BEING_TESTED + "/?filter-by-user=true",
+            ENDPOINT_BEING_TESTED + "?filter-by-user=true",
             waCaseworkerCredentials.getHeaders()
         );
         result.then().assertThat()
@@ -73,7 +73,7 @@ public class GetWorkTypesControllerTest extends SpringBootFunctionalBaseTest {
         common.setupWAOrganisationalRoleAssignment(waCaseworkerCredentials.getHeaders(), "tribunal-caseworker");
 
         Response result = restApiActions.get(
-            ENDPOINT_BEING_TESTED + "/?filter-by-user=true",
+            ENDPOINT_BEING_TESTED + "?filter-by-user=true",
             waCaseworkerCredentials.getHeaders()
         );
         result.then().assertThat()
@@ -92,7 +92,7 @@ public class GetWorkTypesControllerTest extends SpringBootFunctionalBaseTest {
         common.setupWAOrganisationalRoleAssignment(waCaseworkerCredentials.getHeaders(), "tribunal-caseworker");
 
         Response result = restApiActions.get(
-            ENDPOINT_BEING_TESTED + "/?filter-by-user=false",
+            ENDPOINT_BEING_TESTED + "?filter-by-user=false",
             waCaseworkerCredentials.getHeaders()
         );
         result.then().assertThat()
@@ -125,7 +125,7 @@ public class GetWorkTypesControllerTest extends SpringBootFunctionalBaseTest {
     public void should_return_a_403_when_the_user_did_not_have_any_roles() {
 
         Response result = restApiActions.get(
-            ENDPOINT_BEING_TESTED + "/?filter-by-user=true",
+            ENDPOINT_BEING_TESTED + "?filter-by-user=true",
             caseworkerCredentials.getHeaders()
         );
